Completed most of the GUI and some more of the Software
Dec 3,2023

There is a "applet" that will open upon running the code. Users can type in input boxes and select input types by navigating through tabs.

Currently there are tabs for a general solver, area, and volume. As well as a button to denote fractions and and a Solve button/

I still need to connect the mathOps class to the GUI and after that I still need to code some other user interaction.